+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-642** - Possibility to redirect to Department page with filter applied.
 - **GM-642** - SearchContext refactor
 
+### Added
+
+- **GM-664** - Added new order by to Product Gallery
+- **GM-91** - Search page integration
+
 ## [0.14.0] - 2022-08-24
 
 ### Added